requirements.txt
# Core dependencies for personal assistant chatbot
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# For natural language processing
spacy>=3.7.0
nltk>=3.8.1

# For behavior learning and analytics
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0

# For integrations
requests>=2.31.0
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
